Fix inaccuracies in custom element construction stack management
Fixes #1238. Usually, the popped stack entry is not the element, but
instead an already-constructed marker. This makes the stack-popping
generic, and also adds a note explaining exactly what could be going on
here.

Also clarifies the case in which the already constructed marker matters.
It turns out that calling super() multiple times is not possible, but
it's still possible to trigger that exception; this adds an example.

Finally, fixes a typo in the HTMLElement() constructor which used the
variable names "instance" and "element" interchangeably.